Добавилось несколько Сервлетов. Появились аннотации Ломбок

+++ src/main/java/com/alexsandrov/architecture_mvc/entity/Flight.java
@@ -20,6 +20,6 @@ public class Flight {
     private String departureAirportCode;
     private LocalDateTime arrivalDate;
     private String arrivalAirportCode;
-    private Long aircraftId;
+    private Integer aircraftId;
     private FlightStatus status;
 }
